% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunc.R
\name{testfunc}
\alias{testfunc}
\title{Square function}
\usage{
testfunc(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of squared components
}
\description{
Takes vector and returns vector of squared components
}
\examples{
x = 1:30; func(x)
}
